Fixed a bug which reduced accuracy on the ext_timestamp, which is the timestamp in nanoseconds and should be timestamp * 1 billion + nanoseconds, but lost accuracy in the calculation due to an intermediate number being interpreted as a float.
(Mind that we _really_ have to check every part of our calculations to
make sure we don't lose precision with these beasts)

+++ hisparc/eventwarehouse.py
@@ -299,7 +299,9 @@ def process_events(events, eventdata, calculateddata, table, traces):
         tablerow['event_id'] = event_id
         tablerow['timestamp'] = timestamp
         tablerow['nanoseconds'] = nanoseconds
-        tablerow['ext_timestamp'] = timestamp * 1e9 + nanoseconds
+        # careful with intermediate typecasting! we lose precision if we
+        # multiply with just 1e9.
+        tablerow['ext_timestamp'] = timestamp * long(1e9) + nanoseconds
 
         # get default values for the eventdata
         data = {}
